fix(perf-opt): medium-tier code-review findings (memset under lock, UB cast, dead checks, MAX_THREADS override)
Four targeted fixes from independent review of the perf-opt branch:

1. libobs/obs-audio-pool.c — move full memset(block, 0, bsz) inside
   the pool lock. Previously only the first sizeof(void*) bytes (the
   embedded freelist pointer) were zeroed under lock; the rest of the
   block was zeroed AFTER mutex release. Closes a (currently
   theoretical) info-leak where a future consumer reading offset >=
   sizeof(void*) before the post-unlock memset returned would see
   uninitialised previous-tenant audio data. Cost: dozens of ns added
   to each allocation; acceptable.

2. libobs/obs-audio-optimized.c — cast pointers to uintptr_t before
   the dst/src overlap-detection comparison. Comparing pointers from
   unrelated allocations via < / >= is undefined behaviour per the C
   standard and triggers UBSan warnings on sanitizer builds. The cast
   form is semantically identical on flat address spaces (x86-64,
   ARM64) and well-defined under integer comparison rules. Removed
   the 'acknowledged UB' comment.

3. frontend/widgets/OBSBasic.cpp — remove dead UNC/device-path
   prefix checks. The broader '\' double-separator check already
   covers '\?\' and '\.\' before the later separate checks
   for those two prefixes can evaluate. Cosmetic cleanup; behaviour
   unchanged.

4. libobs/obs-audio-threaded.c — gate MAX_THREADS = 16u behind
   #ifndef OBS_AUDIO_MAX_THREADS so downstream packagers can override
   via -DOBS_AUDIO_MAX_THREADS=N from CMake/build flags without
   patching the source. Existing usages unchanged.

+++ libobs/obs-audio-optimized.c
@@ -221,15 +221,16 @@ void copy_video_plane_optimized(uint8_t *dst, const uint8_t *src,
 	}
 
 	/* Detect overlap and fall back to memmove (line-by-line). The SIMD
-	 * paths below assume non-overlapping buffers; comparing pointers across
-	 * unrelated allocations is technically UB but in practice safe on the
-	 * flat address spaces this code targets. We check the full-extent byte
-	 * ranges so legitimate back-to-back single-line buffers don't trip the
-	 * fallback. */
+	 * paths below assume non-overlapping buffers. We compare via uintptr_t
+	 * casts rather than raw pointer relational ops, which is well-defined
+	 * across unrelated allocations on flat address spaces (x86-64, ARM64).
+	 * Check the full-extent byte ranges so legitimate back-to-back
+	 * single-line buffers don't trip the fallback. */
 	{
 		const uint8_t *src_end = src + (size_t)(height - 1) * src_stride + width;
 		const uint8_t *dst_end = dst + (size_t)(height - 1) * dst_stride + width;
-		const bool overlap = !(dst >= src_end || src >= dst_end);
+		const bool overlap = !((uintptr_t)dst >= (uintptr_t)src_end ||
+				       (uintptr_t)src >= (uintptr_t)dst_end);
 		if (overlap) {
 			for (uint32_t y = 0; y < height; y++) {
 				const uint8_t *src_line = src + (size_t)y * src_stride;
